Move write_bytes()/read_bytes() into a separate file
