+++ src/ztype.rs
@@ -15,6 +15,8 @@ pub use self::bits_encode::write_u8;
 
 pub use self::bits_encode::write_signed_bits;
 pub use self::bits_encode::write_unsigned_bits;
+pub use self::varuint_encode::write_varsize;
+pub use self::varuint_encode::write_varuint;
 
 pub use self::bits_decode::read_i16;
 pub use self::bits_decode::read_i32;
@@ -56,7 +58,10 @@ pub mod array;
 pub mod array_traits;
 pub mod bit_alignment;
 
+pub mod bytes_type;
 pub mod extern_type;
 pub mod float_decode;
 pub mod float_encode;
+pub mod reader;
 pub mod traits;
+pub mod writer;
